solana_loader_v4_interface
Module instruction

Instructions for the v4 built-in loader program.

Enums

LoaderV4Instruction

Functions

copy

Returns the instructions required to copy a chunk of program data.

create_buffer

Returns the instructions required to initialize a program/buffer account.

deploy

Returns the instructions required to deploy a program.

deploy_from_source

Returns the instructions required to deploy a program using a buffer.

finalize

Returns the instructions required to finalize program.

is_copy_instruction

is_deploy_instruction

is_finalize_instruction

is_retract_instruction

is_set_program_length_instruction

is_transfer_authority_instruction

is_write_instruction

retract

Returns the instructions required to retract a program.

set_program_length

Returns the instructions required to set the length of the program account.

transfer_authority

Returns the instructions required to transfer authority over a program.

write

Returns the instructions required to write a chunk of program data to a buffer account.